+++ tests/libs/fixtures/__init__.py
@@ -1,2 +1,3 @@
+from .cli import *  # noqa
 from .client import *  # noqa
 from .web3 import *  # type: ignore # noqa
